# src/lib/swagger/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OrbitABM API",
    "description": "RESTful API for OrbitABM - Account-Based Marketing Campaign Intelligence Platform",
    "version": "1.0.0",
    "contact": {
      "name": "OrbitABM Support",
      "email": "support@orbitabm.com"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://orbitabm.com/license"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Development server"
    },
    {
      "url": "https://your-domain.com/api",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/document-templates": {
      "get": {
        "tags": ["Document Templates"],
        "summary": "List document templates",
        "description": "Retrieve a list of document templates with optional filtering and pagination",
        "parameters": [
          {
            "$ref": "#/components/parameters/OrganizationId"
          },
          {
            "name": "document_type",
            "in": "query",
            "description": "Filter by document type",
            "schema": {
              "type": "string",
              "enum": ["prospect_research", "campaign_sequence", "competitive_analysis", "audit_report", "proposal"]
            }
          },
          {
            "name": "vertical_id",
            "in": "query",
            "description": "Filter by vertical UUID",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "is_active",
            "in": "query",
            "description": "Filter by active status",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Offset"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/DocumentTemplate"
                      }
                    },
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "count": {
                      "type": "integer",
                      "nullable": true
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "post": {
        "tags": ["Document Templates"],
        "summary": "Create document template",
        "description": "Create a new document template",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentTemplateCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Document template created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/DocumentTemplate"
                    },
                    "success": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
    "/email-templates": {
      "get": {
        "tags": ["Email Templates"],
        "summary": "List email templates",
        "description": "Retrieve a list of email templates with optional filtering and pagination",
        "parameters": [
          {
            "$ref": "#/components/parameters/OrganizationId"
          },
          {
            "name": "playbook_step_id",
            "in": "query",
            "description": "Filter by playbook step UUID",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "target_contact_role",
            "in": "query",
            "description": "Filter by target contact role",
            "schema": {
              "type": "string",
              "enum": ["economic_buyer", "technical_buyer", "brand_buyer", "champion", "any"]
            }
          },
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Offset"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/EmailTemplate"
                      }
                    },
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "count": {
                      "type": "integer",
                      "nullable": true
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "post": {
        "tags": ["Email Templates"],
        "summary": "Create email template",
        "description": "Create a new email template",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailTemplateCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Email template created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/EmailTemplate"
                    },
                    "success": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/email-templates/{id}": {
      "get": {
        "tags": ["Email Templates"],
        "summary": "Get email template by ID",
        "description": "Retrieve a single email template by its UUID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Email template UUID",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/EmailTemplate"
                    },
                    "success": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "patch": {
        "tags": ["Email Templates"],
        "summary": "Update email template",
        "description": "Update an existing email template (partial update)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Email template UUID",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailTemplateUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Email template updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/EmailTemplate"
                    },
                    "success": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "delete": {
        "tags": ["Email Templates"],
        "summary": "Delete email template",
        "description": "Hard delete an email template",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Email template UUID",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Email template deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "deleted": {
                          "type": "boolean",
                          "example": true
                        }
                      }
                    },
                    "success": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "DocumentTemplate": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid", "description": "Document template UUID" },
          "organization_id": { "type": "string", "format": "uuid", "description": "Organization UUID" },
          "name": { "type": "string", "description": "Template name", "example": "HVAC Prospect Research" },
          "document_type": { "type": "string", "enum": ["prospect_research", "campaign_sequence", "competitive_analysis", "audit_report", "proposal"], "description": "Document type" },
          "vertical_id": { "type": "string", "format": "uuid", "nullable": true, "description": "Vertical UUID" },
          "template_structure": { "type": "object", "description": "Template structure definition (JSON)" },
          "version": { "type": "integer", "description": "Template version", "example": 1 },
          "is_active": { "type": "boolean", "description": "Whether template is active", "example": true },
          "created_at": { "type": "string", "format": "date-time", "description": "Creation timestamp" },
          "updated_at": { "type": "string", "format": "date-time", "description": "Last update timestamp" },
          "deleted_at": { "type": "string", "format": "date-time", "nullable": true, "description": "Deletion timestamp" }
        },
        "required": ["id", "organization_id", "name", "document_type", "template_structure"]
      },
      "DocumentTemplateCreate": {
        "type": "object",
        "properties": {
          "organization_id": { "type": "string", "format": "uuid", "description": "Organization UUID" },
          "name": { "type": "string", "description": "Template name", "example": "HVAC Prospect Research" },
          "document_type": { "type": "string", "enum": ["prospect_research", "campaign_sequence", "competitive_analysis", "audit_report", "proposal"], "description": "Document type" },
          "vertical_id": { "type": "string", "format": "uuid", "description": "Vertical UUID" },
          "template_structure": { "type": "object", "description": "Template structure definition (JSON)" },
          "version": { "type": "integer", "default": 1, "description": "Template version" },
          "is_active": { "type": "boolean", "default": true, "description": "Whether template is active" }
        },
        "required": ["organization_id", "name", "document_type", "template_structure"]
      },
...
      "EmailTemplate": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid", "description": "Email template UUID" },
          "organization_id": { "type": "string", "format": "uuid", "description": "Organization UUID" },
          "playbook_step_id": { "type": "string", "format": "uuid", "nullable": true, "description": "Playbook step UUID" },
          "name": { "type": "string", "description": "Template name", "example": "Initial Outreach - Owner" },
          "subject_line": { "type": "string", "description": "Email subject line", "example": "Quick question about your marketing" },
          "subject_line_alt": { "type": "string", "nullable": true, "description": "Alternative subject line" },
          "body": { "type": "string", "description": "Email body content" },
          "target_contact_role": { "type": "string", "enum": ["economic_buyer", "technical_buyer", "brand_buyer", "champion", "any"], "nullable": true, "description": "Target contact role" },
          "merge_fields_required": { "type": "array", "items": { "type": "string" }, "nullable": true, "description": "Required merge fields" },
          "notes": { "type": "string", "nullable": true, "description": "Additional notes" },
          "created_at": { "type": "string", "format": "date-time", "description": "Creation timestamp" },
          "updated_at": { "type": "string", "format": "date-time", "description": "Last update timestamp" }
        },
        "required": ["id", "organization_id", "name", "subject_line", "body"]
      },
      "EmailTemplateCreate": {
        "type": "object",
        "properties": {
          "organization_id": { "type": "string", "format": "uuid", "description": "Organization UUID" },
          "playbook_step_id": { "type": "string", "format": "uuid", "description": "Playbook step UUID" },
          "name": { "type": "string", "description": "Template name", "example": "Initial Outreach - Owner" },
          "subject_line": { "type": "string", "description": "Email subject line" },
          "subject_line_alt": { "type": "string", "description": "Alternative subject line" },
          "body": { "type": "string", "description": "Email body content" },
          "target_contact_role": { "type": "string", "enum": ["economic_buyer", "technical_buyer", "brand_buyer", "champion", "any"], "description": "Target contact role" },
          "merge_fields_required": { "type": "array", "items": { "type": "string" }, "description": "Required merge fields" },
          "notes": { "type": "string", "description": "Additional notes" }
        },
        "required": ["organization_id", "name", "subject_line", "body"]
      },
      "EmailTemplateUpdate": {
        "type": "object",
        "properties": {
          "playbook_step_id": { "type": "string", "format": "uuid", "nullable": true, "description": "Playbook step UUID" },
          "name": { "type": "string", "description": "Template name" },
          "subject_line": { "type": "string", "description": "Email subject line" },
          "subject_line_alt": { "type": "string", "nullable": true, "description": "Alternative subject line" },
          "body": { "type": "string", "description": "Email body content" },
          "target_contact_role": { "type": "string", "enum": ["economic_buyer", "technical_buyer", "brand_buyer", "champion", "any"], "nullable": true, "description": "Target contact role" },
          "merge_fields_required": { "type": "array", "items": { "type": "string" }, "nullable": true, "description": "Required merge fields" },
          "notes": { "type": "string", "nullable": true, "description": "Additional notes" }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable error message",
            "example": "Company with this name already exists"
          },
          "code": {
            "type": "string",
            "description": "Error code",
            "enum": ["VALIDATION_ERROR", "NOT_FOUND", "DUPLICATE_ENTRY", "FORBIDDEN", "DATABASE_ERROR", "SERVER_ERROR", "NETWORK_ERROR"],
            "example": "DUPLICATE_ENTRY"
          },
          "success": {
            "type": "boolean",
            "example": false
          },
          "details": {
            "type": "object",
            "nullable": true,
            "description": "Additional error details",
            "example": {
              "field": "name",
              "value": "Acme HVAC"
            }
          }
        },
        "required": ["error", "code", "success"]
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Bad Request - Invalid input data",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "Organization ID is required",
              "code": "VALIDATION_ERROR",
              "success": false
            }
          }
        }
      },
      "NotFound": {
        "description": "Not Found - Resource does not exist",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "Company not found",
              "code": "NOT_FOUND",
              "success": false
            }
          }
        }
      },
...
      "InternalServerError": {
        "description": "Internal Server Error - Unexpected server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "Internal server error",
              "code": "SERVER_ERROR",
              "success": false
            }
          }
        }
      }
    },
    "parameters": {
      "OrganizationId": {
        "name": "organization_id",
        "in": "query",
        "required": true,
        "description": "Organization UUID",
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
      "Limit": {
        "name": "limit",
        "in": "query",
        "description": "Number of results per page",
        "schema": {
          "type": "integer",
          "default": 100,
          "minimum": 1,
          "maximum": 1000
        }
      },
      "Offset": {
        "name": "offset",
        "in": "query",
        "description": "Number of results to skip",
        "schema": {
          "type": "integer",
          "default": 0,
          "minimum": 0
        }
      }
    }
  }
}